refactor: change modal
