+++ react-rss/src/pages/FormsPage/components/Modal/Modal.module.scss
@@ -1,0 +1,21 @@
+.showModal {
+  background-color: #fff;
+  border-radius: 20px;
+  width: 520px;
+  height: 575px;
+  z-index: 7;
+  display: flex;
+  margin: 0 auto;
+  position: absolute;
+  top: 50%;
+  right: 50%;
+  transform: translate(50%, -50%);
+  font-size: 75px;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+}
+
+.hideModal {
+  display: none;
+}
