+++ react-rss/src/pages/FormsPage/components/Modal/index.ts
@@ -1,2 +1,1 @@
 export { default } from './Modal'
-export * from './hooks'
