+++ react-rss/src/pages/HomePage/HomePage.tsx
@@ -1,10 +1,9 @@
 import { FC } from 'react'
 
 import SearchInput from '../../components/SearchInput'
-// import mainImage from '../../assets/images/clotes.jpg'
 import Layout from '../../components/Layout'
 
-import ModalProductInfo from './components/ModalProductInfo'
+import ModalProductInfo from '../../components/Modal'
 import { useHomePage, getInfoproduct } from './hooks'
 import CardsList from './components/CardsList'
 
@@ -17,19 +16,9 @@ const HomePage: FC = () => {
   const { modalVisible, showInfoProduct, selectCharacter, closeModal } =
     getInfoproduct()
 
-  //! REFACTOR ALL CODE!!!!
-
   return (
     <div data-testid="homeContainer">
       <Layout>
-        {/* <div className={styles.wrapperMain}>
-          <div className={styles.shadow}></div>
-          <img src={mainImage} className={styles.mainImage}></img>
-          <div className={styles.mainTitle}>
-            bajor
-            <br /> -<br /> choose the best
-          </div>
-        </div> */}
         <SearchInput
           value={searchString || ''}
           onChange={(event) => onChangeSearch(event?.target.value || '')} //? Remove div wrapper
