--- react-rss/src/pages/HomePage/components/ModalProductInfo/index.ts
@@ -1,2 +1,0 @@
-export { default } from './ModalProductInfo'
-export * from './hooks'
